Step4 요금 조회 PR 요청드립니다!

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -46,7 +46,13 @@ public class FavoriteService {
     }
 
     @Transactional
-    public void deleteFavorite(long id) {
+    public void deleteFavorite(LoginMember loginMember, long id) {
+        Favorite favorite = findFavoriteById(id);
+        favorite.validateOwner(loginMember);
         favoriteRepository.deleteById(id);
+    }
+
+    private Favorite findFavoriteById(long id) {
+        return favoriteRepository.findById(id).orElseThrow(() -> new NotFoundException("조회되는 즐겨찾기가 없습니다."));
     }
 }
